Move diagnostics to separate dir
